hw: acpi: Factor some i386 ACPI code out
We move some structure into the architecture agnostic
include/hw/acpi/acpi.h header, and we move the ACPI link
API into hw/acpi/link.c

We need to do so in preparation for the generic HW reduced ACPI
implementation.
