+++ hw/i386/acpi-build.h
@@ -2,7 +2,7 @@
 #ifndef HW_I386_ACPI_BUILD_H
 #define HW_I386_ACPI_BUILD_H
 
-#include "hw/i386/acpi.h"
+#include "hw/acpi/acpi.h"
 
 void acpi_setup(MachineState *machine, AcpiConfiguration *conf);
